Restructure to make more upstreamable
I'd potentially like to use this PDK as a guinea pig for upstreaming
the Cedar support to an official pdk package. As discussed, it seems
unlikely that we can do that for openpdks for the forseeable future,
since that's managed as a git-repo + pile-of-patches. However,
OpenROAD manages ASAP7 as a github repo (even though they are not
the authors), so it seems possible we could upstream it there. I
don't think we're quite ready for that conversation yet, but to that
end, at lest:

- Move the julia files into a `cedar` subdirectory
- rm `authors` field from the Project.toml, since the Project.toml
  applies to the whole package in principle and we didn't write the PDK

I'd also, ideally like to get rid of the `CMC` dependency (cedar
should load that internally when it sees as appropriate level),
but that's fine for now.